+++ PeriodicTopicBehaviors/src/main/java/edu/ntnu/app/psta/PstaPattern.java
@@ -3,12 +3,12 @@ package edu.ntnu.app.psta;
 import java.util.TreeSet;
 
 public class PstaPattern {
-    private final Float period;
+    private final double period;
     private final long offset;
     private final int themeIndex;
     private final TreeSet<Integer> locationTrajectory;
 
-    public PstaPattern(Float period, long offset, int themeIndex) {
+    public PstaPattern(double period, long offset, int themeIndex) {
         this.period = period;
         this.offset = offset;
         this.themeIndex = themeIndex;
@@ -19,7 +19,7 @@ public class PstaPattern {
         locationTrajectory.add(l);
     }
 
-    public Float getPeriod() {
+    public double getPeriod() {
         return period;
     }
